chore(code): merge all useDoc hooks
